# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "中医药知识库 API 服务 —— 方剂、单味药、对药、论文、分子信息、名家经验管理",
        "title": "AI Medical Agent API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/agent/sessions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取当前用户的所有 Agent 对话会话",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agent-chat"
                ],
                "summary": "获取对话会话列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/medicalagent_internal_model.AgentSession"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "为用户创建一个新的 Agent 对话会话",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agent-chat"
                ],
                "summary": "创建对话会话",
                "parameters": [
                    {
                        "description": "会话信息",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.createAgentSessionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "data": {
                                    "$ref": "#/definitions/medicalagent_internal_model.AgentSession"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/agent/sessions/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取指定会话的详情及所有历史消息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agent-chat"
                ],
                "summary": "获取会话详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "会话 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "data": {
                                    "type": "object",
                                    "properties": {
                                        "messages": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/medicalagent_internal_model.AgentChatMessage"
                                            }
                                        },
                                        "session": {
                                            "$ref": "#/definitions/medicalagent_internal_model.AgentSession"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/agent/sessions/{id}/messages": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "在指定会话中发送消息，返回生成的 Turn ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agent-chat"
                ],
                "summary": "发送对话消息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "会话 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "消息内容",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler.sendAgentMessageRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "data": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string"
                                        },
                                        "turn_id": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/agent/sessions/{id}/turns/{turn_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取指定对话轮次（Turn）的当前状态 and 分析结果",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agent-chat"
                ],
                "summary": "获取对话轮次结果",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "会话 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "轮次 ID",
                        "name": "turn_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "data": {
                                    "$ref": "#/definitions/medicalagent_internal_model.AgentChatTurnResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/agent/sessions/{id}/turns/{turn_id}/stream": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "通过 SSE 流式获取轮次的执行进度、事件以及最终结果。\n事件帧格式为 Data: {sequence: int64, ...}。\n事件类型包括:\n1. \"execution\": 包含中间执行步骤的事件帧。\n2. \"turn\": 包含最终状态 (done/failed) 和完整结果的结束帧。\n当收到 status 为 \"done\" 或 \"failed\" 的 \"turn\" 事件，或连接断开时，流式传输结束。",
                "produces": [
                    "text/event-stream"
                ],
                "tags": [
                    "agent-chat"
                ],
                "summary": "流式获取轮次进度 (SSE)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "会话 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "轮次 ID",
                        "name": "turn_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SSE stream",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_handler.createAgentSessionRequest": {
            "type": "object",
            "properties": {
                "title": {
                    "type": "string",
                    "example": "关于麻黄的药理咨询"
                }
            }
        },
...
        "internal_handler.sendAgentMessageRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string",
                    "example": "请分析麻黄在宣肺平喘方面的科学原理"
                }
            }
        },
...
        "medicalagent_internal_model.AgentChatMessage": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "role": {
                    "type": "string"
                },
                "session_id": {
                    "type": "integer"
                },
                "turn_id": {
                    "type": "string"
                }
            }
        },
        "medicalagent_internal_model.AgentChatTurnResponse": {
            "type": "object",
            "properties": {
                "assistant_message": {
                    "$ref": "#/definitions/medicalagent_internal_model.AgentChatMessage"
                },
                "created_at": {
                    "type": "string"
                },
                "error_message": {
                    "type": "string"
                },
                "events": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "session_id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "turn_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "medicalagent_internal_model.AgentSession": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "AgentTokenAuth": {
            "type": "apiKey",
            "name": "X-Agent-Token",
            "in": "header"
        },
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}